Refactor: Implement MVVM-like structure and UI improvements
This major refactoring reorganizes the application towards an MVVM-like
pattern and addresses several UI and project management requirements.

Key changes include:

1.  **File Structure:**
    *   Introduced `models`, `stores`, `widgets`, and `views` directories.
    *   Split `ui_main.py` into `mainwindow.py`, `models/project.py`,
        `widgets/project_editor.py`, `widgets/results_viewer.py`,
        `widgets/sidebar.py`, and `views/workspace.py`.

2.  **Project Management (`ProjectStore`):**
    *   `ProjectStore` (`stores/project_store.py`) now manages project
        CRUD operations and persistence.
    *   Projects are stored in `~/.config/regulens-ai/projects.json`.
    *   `CompareProject` (`models/project.py`) updated with `updated` and
        `deleted` signals for better synchronization.

3.  **UI Flow and Components:**
    *   `MainWindow` (`mainwindow.py`) is now a lean application shell.
        Line count reduced from ~500 to ~168.
    *   An `IntroPage` (`widgets/intro_page.py`) is shown on startup.
    *   `Workspace` (`views/workspace.py`) is the main view after "Get Started",
        containing the sidebar and a stacked widget for editors/viewers.
    *   `Sidebar` (`widgets/sidebar.py`) lists projects, allows selection,
        creation, and features a new toggle mechanism that resizes the
        splitter pane instead of hiding the widget.
    *   `ProjectEditor` and `ResultsViewer` are decoupled and communicate
        via signals within the `Workspace`.

4.  **Cleanup:**
    *   Removed obsolete methods and UI construction logic from `MainWindow`.
    *   Updated `app/__init__.py`.
    *   Confirmed `SettingsDialog` usage and addressed potential duplicate files.

5.  **Testing:**
    *   The test suite passes after I addressed test setup issues related
        to the refactoring.

These changes fulfill the requirements you outlined, leading to a more modular, maintainable, and user-friendly application structure.
