+++ app/__init__.py
@@ -1,7 +1,7 @@
 """Regulens-AI application package."""
 
-try:  # optional dependency for tests
-    from .ui_main import MainWindow
+try:
+    from .mainwindow import MainWindow
 except Exception:  # pragma: no cover - optional GUI
     MainWindow = None  # type: ignore
